Refactor 테스트 코드에서 반복되어 사용되는 테스트용 데이터를 fixture로 분리한다
