+++ __fixtures__/restaurants.js
@@ -1,0 +1,20 @@
+const restaurants = [{
+  id: 1,
+  name: '시카고피자',
+  category: '양식',
+  address: '이태원동',
+},
+{
+  id: 2,
+  name: '마녀주방',
+  category: '한식',
+  address: '서울시 강남구',
+},
+{
+  id: 3,
+  name: '키와미',
+  category: '일식',
+  address: '분당구 정자동',
+}];
+
+export default restaurants;
