+++ src/List.test.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 
 import { render, screen } from '@testing-library/react';
 
+import restaurantsFixtures from '../__fixtures__/restaurants';
+
 import List from './List';
 
 function renderList(restaurants) {
@@ -33,24 +35,7 @@ describe('<List />', () => {
 
   context('with restaurants', () => {
     beforeEach(() => {
-      restaurants = [{
-        id: 1,
-        name: '시카고피자',
-        category: '양식',
-        address: '이태원동',
-      },
-      {
-        id: 2,
-        name: '마녀주방',
-        category: '한식',
-        address: '서울시 강남구',
-      },
-      {
-        id: 3,
-        name: '키와미',
-        category: '일식',
-        address: '분당구 정자동',
-      }];
+      restaurants = restaurantsFixtures;
     });
 
     it('print the restaurant items', () => {
